+++ zukeUI/assets/zukeUI.ui
@@ -2,6 +2,9 @@
 <ui version="4.0">
  <class>Form</class>
  <widget class="QWidget" name="Form">
+  <property name="windowModality">
+   <enum>Qt::ApplicationModal</enum>
+  </property>
   <property name="geometry">
    <rect>
     <x>0</x>
@@ -9,6 +12,9 @@
     <width>569</width>
     <height>708</height>
    </rect>
+  </property>
+  <property name="contextMenuPolicy">
+   <enum>Qt::DefaultContextMenu</enum>
   </property>
   <property name="windowTitle">
    <string>ZukeUI</string>
@@ -146,14 +152,17 @@
   <widget class="QSlider" name="volume_slider">
    <property name="geometry">
     <rect>
-     <x>50</x>
+     <x>130</x>
      <y>410</y>
-     <width>461</width>
+     <width>381</width>
      <height>20</height>
     </rect>
    </property>
    <property name="accessibleName">
     <string/>
+   </property>
+   <property name="tracking">
+    <bool>false</bool>
    </property>
    <property name="orientation">
     <enum>Qt::Horizontal</enum>
@@ -162,9 +171,9 @@
   <widget class="QSlider" name="seek_slider">
    <property name="geometry">
     <rect>
-     <x>50</x>
+     <x>130</x>
      <y>440</y>
-     <width>461</width>
+     <width>381</width>
      <height>20</height>
     </rect>
    </property>
@@ -174,6 +183,9 @@
    <property name="maximum">
     <number>100</number>
    </property>
+   <property name="tracking">
+    <bool>false</bool>
+   </property>
    <property name="orientation">
     <enum>Qt::Horizontal</enum>
    </property>
@@ -181,9 +193,9 @@
   <widget class="QLabel" name="track_sender">
    <property name="geometry">
     <rect>
-     <x>50</x>
+     <x>142</x>
      <y>380</y>
-     <width>473</width>
+     <width>381</width>
      <height>21</height>
     </rect>
    </property>
@@ -229,6 +241,54 @@
     <set>Qt::AlignCenter</set>
    </property>
   </widget>
+  <widget class="QLabel" name="volume_label">
+   <property name="geometry">
+    <rect>
+     <x>65</x>
+     <y>410</y>
+     <width>61</width>
+     <height>20</height>
+    </rect>
+   </property>
+   <property name="text">
+    <string>Volume:</string>
+   </property>
+   <property name="alignment">
+    <set>Qt::AlignLeading|Qt::AlignLeft|Qt::AlignVCenter</set>
+   </property>
+  </widget>
+  <widget class="QLabel" name="label_2">
+   <property name="geometry">
+    <rect>
+     <x>65</x>
+     <y>440</y>
+     <width>61</width>
+     <height>20</height>
+    </rect>
+   </property>
+   <property name="text">
+    <string>Seek:</string>
+   </property>
+   <property name="alignment">
+    <set>Qt::AlignLeading|Qt::AlignLeft|Qt::AlignVCenter</set>
+   </property>
+  </widget>
+  <widget class="QLabel" name="sentby_label">
+   <property name="geometry">
+    <rect>
+     <x>65</x>
+     <y>380</y>
+     <width>61</width>
+     <height>21</height>
+    </rect>
+   </property>
+   <property name="text">
+    <string>Sent by:</string>
+   </property>
+   <property name="alignment">
+    <set>Qt::AlignLeading|Qt::AlignLeft|Qt::AlignVCenter</set>
+   </property>
+  </widget>
   <zorder>frame_2</zorder>
   <zorder>widget</zorder>
   <zorder>track_title</zorder>
@@ -236,6 +296,9 @@
   <zorder>seek_slider</zorder>
   <zorder>track_sender</zorder>
   <zorder>error_message</zorder>
+  <zorder>volume_label</zorder>
+  <zorder>label_2</zorder>
+  <zorder>sentby_label</zorder>
  </widget>
  <resources/>
  <connections/>
