+++ zukeUI/driver/utils/connection/request_managing.py
@@ -13,17 +13,14 @@ class RequestManager(QtCore.QThread):
         QtCore.QThread.__init__(self)
         self._config_manager = config_manager
         self._requests = []
+        self._refreshing = False
+        self._callback = None
+        self._control = None
 
     def send_request(self, request: ZukeRequest, response_callback=None):
         request = ZukeRequestThread(request, response_callback)
         request.start()
         self._requests.append(request)
-
-    def run(self):
-        pass
-
-
-class UrlSendingManager(RequestManager):
 
     def send_track(self, url, response_callback=None, message: str=None):
 
@@ -42,9 +39,6 @@ class UrlSendingManager(RequestManager):
             response_callback
         )
 
-
-class VolumeSetterManager(RequestManager):
-
     def set_volume(self, volume_value, response_callback=None):
         self.send_request(
             ZukeRequest(
@@ -57,9 +51,6 @@ class VolumeSetterManager(RequestManager):
             ),
             response_callback
         )
-
-
-class SeekSetterManager(RequestManager):
 
     def set_seek(self, seek_value, response_callback=None):
         self.send_request(
@@ -74,8 +65,6 @@ class SeekSetterManager(RequestManager):
             response_callback
         )
 
-
-class PlayPauseManager(RequestManager):
 
     manager_id = "playpause_manager"
     response_signal = stop_signal = QtCore.pyqtSignal(dict, name=manager_id)
@@ -93,33 +82,25 @@ class PlayPauseManager(RequestManager):
             response_callback
         )
 
-
-class RefreshManager(RequestManager):
-
-    manager_id = "refresh_manager"
-    response_signal = stop_signal = QtCore.pyqtSignal(dict, name=manager_id)
-
-    def __init__(self, zcm: ZukeConfigManager):
-        super().__init__(zcm)
-        self.__control = None
-        self._refreshing = False
+    def get_control(self, callback=None):
+        self.send_request(
+            ZukeRequest(
+                self._config_manager.zuke_ip,
+                self._config_manager.zuke_port,
+                "GET",
+                "/player/control"
+            ),
+            callback
+        )
 
     def run(self):
         while self._refreshing:
-                self.send_request(
-                    ZukeRequest(
-                        self._config_manager.zuke_ip,
-                        self._config_manager.zuke_port,
-                        "GET",
-                        "/player/control"
-                    ),
-                    self._callback
-                )
+                self.get_control(self._callback)
                 time.sleep(1)
 
-    @property
-    def control(self):
-        return self.__control
+    def _refresh_callback(self, response):
+        self._control = response
+        self._callback(response)
 
     def start_refreshing(self, callback=None):
         if not self._refreshing:
@@ -132,44 +113,6 @@ class RefreshManager(RequestManager):
             self._refreshing = False
             self.quit()
 
-
-class ZukeUiManager:
-
-    def __init__(
-            self,
-            usm: UrlSendingManager,
-            vsm: VolumeSetterManager,
-            ssm: SeekSetterManager,
-            ppm: PlayPauseManager,
-            rm: RefreshManager
-    ):
-        self.prev_control = None
-        self.__usm = usm
-        self.__vsm = vsm
-        self.__ssm = ssm
-        self.__ppm = ppm
-        self.__rm = rm
-
-    def send_url(self, url: str, message: str, callback=None):
-        if url:
-            self.__usm.send_track(url, callback, message)
-
-    def set_volume(self, volume, callback=None):
-        self.__vsm.set_volume(volume, callback)
-
-    def set_seek(self, seek, callback=None):
-        self.__ssm.set_seek(seek, callback)
-
-    def play_or_pause(self, play_or_pause, callback=None):
-        self.__ppm.play_or_pause(play_or_pause, callback)
-
-    def stop_refreshing(self):
-        self.__rm.stop_refreshing()
-
-    def start_refreshing(self, ui_callback):
-        self.__rm.start_refreshing(ui_callback)
-
     @property
     def control(self):
-        return self.__rm.control
-
+        return self._control
