+++ zukeUI/runner.sh
@@ -1,3 +1,8 @@
 #!/bin/bash
 
-python3 ./ui/zukeui.py ./assets/
+BASEDIR=$(dirname "$0")
+
+if [[ *"zukeUI"* != $PYTHONPATH ]]; then
+    export PYTHONPATH=${BASEDIR}
+fi
+python3 ./${BASEDIR}/ui/zukeui.py ./${BASEDIR}/assets/
